refactor ClassDiagramGenerator
