+++ src/PlantUmlClassDiagramGenerator.Library/ClassDiagramGenerator/ClassDiagramGenerator.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text.RegularExpressions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -10,7 +8,7 @@ using PlantUmlClassDiagramGenerator.Attributes;
 
 namespace PlantUmlClassDiagramGenerator.Library.ClassDiagramGenerator;
 
-public class ClassDiagramGenerator(
+public partial class ClassDiagramGenerator(
     TextWriter writer,
     string indent,
     Accessibilities ignoreMemberAccessibilities = Accessibilities.None,
@@ -51,272 +49,6 @@ public class ClassDiagramGenerator(
             GenerateRelationships();
     }
 
-    public override void VisitFileScopedNamespaceDeclaration(FileScopedNamespaceDeclarationSyntax node)
-    {
-        if (this.addPackageTags)
-            VisitFileScopedNamespaceDeclaration(node, () => base.VisitFileScopedNamespaceDeclaration(node));
-        else
-            base.VisitFileScopedNamespaceDeclaration(node);
-    }
-    
-    public override void VisitNamespaceDeclaration(NamespaceDeclarationSyntax node)
-    {
-        if (this.addPackageTags)
-            VisitNamespaceDeclaration(node, () => base.VisitNamespaceDeclaration(node));
-        else
-            base.VisitNamespaceDeclaration(node);
-    }
-
-    public override void VisitInterfaceDeclaration(InterfaceDeclarationSyntax node)
-    {
-        VisitTypeDeclaration(node, () => base.VisitInterfaceDeclaration(node));
-    }
-
-    public override void VisitClassDeclaration(ClassDeclarationSyntax node)
-    {
-        VisitTypeDeclaration(node, () => base.VisitClassDeclaration(node));
-    }
-
-    public override void VisitRecordDeclaration(RecordDeclarationSyntax node)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        relationships.AddInnerclassRelationFrom(node);
-        relationships.AddInheritanceFrom(node);
-        var modifiers = GetTypeModifiersText(node.Modifiers);
-        var abstractKeyword = (node.Modifiers.Any(SyntaxKind.AbstractKeyword) ? "abstract " : "");
-
-        var typeName = TypeNameText.From(node);
-        var name = typeName.Identifier;
-        var typeParam = typeName.TypeArguments;
-        var type = $"{name}{typeParam}";
-        var typeParams = typeParam.TrimStart('<').TrimEnd('>').Split([','], StringSplitOptions.RemoveEmptyEntries);
-        types.Add(name);
-
-        var typeKeyword = (node.Kind() == SyntaxKind.RecordStructDeclaration) ? "struct" : "class";
-        WriteLine($"{abstractKeyword}{typeKeyword} {type} {modifiers}<<record>> {{");
-
-        nestingDepth++;
-        var parameters = node.ParameterList?.Parameters ?? Enumerable.Empty<ParameterSyntax>();
-        foreach (var parameter in parameters)
-        {
-            VisitRecordParameter(node, type, typeParams, parameter);
-        }
-        base.VisitRecordDeclaration(node);
-        nestingDepth--;
-
-        WriteLine("}");
-    }
-
-    private void VisitRecordParameter(RecordDeclarationSyntax node, string type, string[] typeParams, ParameterSyntax parameter)
-    {
-        var parameterType = parameter.Type;
-        var isTypeParameterProp = typeParams.Contains(parameterType.ToString());
-        var associationAttrSyntax = parameter.AttributeLists.GetAssociationAttributeSyntax();
-        if (associationAttrSyntax is not null)
-        {
-            var associationAttr = CreateAssociationAttribute(associationAttrSyntax);
-            relationships.AddAssociationFrom(node, parameter, associationAttr);
-        }
-        else if (!createAssociation
-          || parameter.AttributeLists.HasIgnoreAssociationAttribute()
-          || parameterType.GetType() == typeof(PredefinedTypeSyntax)
-          || parameterType.GetType() == typeof(NullableTypeSyntax)
-          || isTypeParameterProp)
-        {
-            // ParameterList-Property: always public
-            var parameterModifiers = "+ ";
-            var parameterName = parameter.Identifier.ToString();
-
-            // ParameterList-Property always have get and init accessor
-            var accessorStr = "<<get>> <<init>>";
-
-            var useLiteralInit = parameter.Default?.Value is not null;
-            var initValue = useLiteralInit
-                ? (" = " + escapeDictionary.Aggregate(parameter.Default.Value.ToString(),
-                    (n, e) => Regex.Replace(n, e.Key, e.Value)))
-                : "";
-            WriteLine($"{parameterModifiers}{parameterName} : {parameterType} {accessorStr}{initValue}");
-        }
-        else
-        {
-            if (type.GetType() == typeof(GenericNameSyntax))
-            {
-                additionalTypeDeclarationNodes.Add(parameterType);
-            }
-            relationships.AddAssociationFrom(parameter, node);
-        }
-    }
-
-    public override void VisitStructDeclaration(StructDeclarationSyntax node)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        relationships.AddInnerclassRelationFrom(node);
-        relationships.AddInheritanceFrom(node);
-
-        var typeName = TypeNameText.From(node);
-        var name = typeName.Identifier;
-        var typeParam = typeName.TypeArguments;
-        var type = $"{name}{typeParam}";
-
-        types.Add(name);
-
-        WriteLine($"struct {type} {{");
-
-        nestingDepth++;
-        base.VisitStructDeclaration(node);
-        nestingDepth--;
-
-        WriteLine("}");
-    }
-
-    public override void VisitEnumDeclaration(EnumDeclarationSyntax node)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        relationships.AddInnerclassRelationFrom(node);
-
-        var type = $"{node.Identifier}";
-
-        types.Add(type);
-
-        WriteLine($"{node.EnumKeyword} {type} {{");
-
-        nestingDepth++;
-        base.VisitEnumDeclaration(node);
-        nestingDepth--;
-
-        WriteLine("}");
-    }
-
-    public override void VisitConstructorDeclaration(ConstructorDeclarationSyntax node)
-    {
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (IsIgnoreMember(node.Modifiers)) { return; }
-        foreach (var parameter in node.ParameterList?.Parameters)
-        {
-            var associationAttrSyntax = parameter.AttributeLists.GetAssociationAttributeSyntax();
-            if (associationAttrSyntax is not null)
-            {
-                var associationAttr = CreateAssociationAttribute(associationAttrSyntax);
-                relationships.AddAssociationFrom(node, parameter, associationAttr);
-            }
-        }
-        var modifiers = GetMemberModifiersText(node.Modifiers,
-            isInterfaceMember: node.Parent.IsKind(SyntaxKind.InterfaceDeclaration));
-        var name = node.Identifier.ToString();
-        var args = node.ParameterList.Parameters.Select(p => $"{p.Identifier}:{p.Type}");
-
-        WriteLine($"{modifiers}{name}({string.Join(", ", args)})");
-    }
-
-    public override void VisitFieldDeclaration(FieldDeclarationSyntax node)
-    {
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (IsIgnoreMember(node.Modifiers)) { return; }
-
-        var modifiers = GetMemberModifiersText(node.Modifiers,
-                isInterfaceMember: node.Parent.IsKind(SyntaxKind.InterfaceDeclaration));
-        var type = node.Declaration.Type;
-        var variables = node.Declaration.Variables;
-        var parentClass = (node.Parent as TypeDeclarationSyntax);
-        var isTypeParameterField = parentClass?.TypeParameterList?.Parameters
-            .Any(t => t.Identifier.Text == type.ToString()) ?? false;
-
-        foreach (var field in variables)
-        {
-            Type fieldType = type.GetType();
-            var associationAttrSyntax = node.AttributeLists.GetAssociationAttributeSyntax();
-            if (associationAttrSyntax is not null)
-            {
-                var associationAttr = CreateAssociationAttribute(associationAttrSyntax);
-                relationships.AddAssociationFrom(node, associationAttr);
-            }
-            else if (!createAssociation
-                || node.AttributeLists.HasIgnoreAssociationAttribute()
-                || fieldType == typeof(PredefinedTypeSyntax)
-                || fieldType == typeof(NullableTypeSyntax)
-                || isTypeParameterField)
-            {
-                var useLiteralInit = field.Initializer?.Value?.Kind().ToString().EndsWith("LiteralExpression") ?? false;
-                var initValue = useLiteralInit
-                    ? (" = " + escapeDictionary.Aggregate(field.Initializer.Value.ToString(),
-                        (f, e) => Regex.Replace(f, e.Key, e.Value)))
-                    : "";
-                WriteLine($"{modifiers}{field.Identifier} : {type}{initValue}");
-            }
-            else
-            {
-                if (fieldType == typeof(GenericNameSyntax))
-                {
-                    additionalTypeDeclarationNodes.Add(type);
-                }
-                relationships.AddAssociationFrom(node, field);
-            }
-        }
-    }
-
-    public override void VisitPropertyDeclaration(PropertyDeclarationSyntax node)
-    {
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (IsIgnoreMember(node.Modifiers)) { return; }
-
-        var type = node.Type;
-
-        var parentClass = (node.Parent as TypeDeclarationSyntax);
-        var isTypeParameterProp = parentClass?.TypeParameterList?.Parameters
-            .Any(t => t.Identifier.Text == type.ToString()) ?? false;
-
-        var typeIgnoringNullable = type is NullableTypeSyntax nullableTypeSyntax ? nullableTypeSyntax.ElementType : type;
-
-        var associationAttrSyntax = node.AttributeLists.GetAssociationAttributeSyntax();
-        if (associationAttrSyntax is not null)
-        {
-            var associationAttr = CreateAssociationAttribute(associationAttrSyntax);
-            relationships.AddAssociationFrom(node, associationAttr);
-        }
-        else if (!createAssociation
-            || node.AttributeLists.HasIgnoreAssociationAttribute()
-            || typeIgnoringNullable is PredefinedTypeSyntax
-            || isTypeParameterProp)
-        {
-            var modifiers = GetMemberModifiersText(node.Modifiers,
-                isInterfaceMember: node.Parent.IsKind(SyntaxKind.InterfaceDeclaration));
-            var name = node.Identifier.ToString();
-            //Property does not have an accessor is an expression-bodied property. (get only)
-            var accessorStr = "<<get>>";
-            if (node.AccessorList != null)
-            {
-                var accessor = node.AccessorList.Accessors
-                    .Where(x => !x.Modifiers.Select(y => y.Kind()).Contains(SyntaxKind.PrivateKeyword))
-                    .Select(x => $"<<{(x.Modifiers.ToString() == "" ? "" : (x.Modifiers.ToString() + " "))}{x.Keyword}>>");
-                accessorStr = string.Join(" ", accessor);
-            }
-            var useLiteralInit = node.Initializer?.Value?.Kind().ToString().EndsWith("LiteralExpression") ?? false;
-            var initValue = useLiteralInit
-                ? (" = " + escapeDictionary.Aggregate(node.Initializer.Value.ToString(),
-                    (n, e) => Regex.Replace(n, e.Key, e.Value)))
-                : "";
-
-            WriteLine($"{modifiers}{name} : {type} {accessorStr}{initValue}");
-        }
-        else
-        {
-            if (type.GetType() == typeof(GenericNameSyntax))
-            {
-                additionalTypeDeclarationNodes.Add(type);
-            }
-            relationships.AddAssociationFrom(node, typeIgnoringNullable);
-        }
-    }
-
     private static PlantUmlAssociationAttribute CreateAssociationAttribute(AttributeSyntax associationAttribute)
     {
         var attributeProps = associationAttribute.ArgumentList.Arguments.Select(arg => new
@@ -334,44 +66,7 @@ public class ClassDiagramGenerator(
         };
     }
 
-    public override void VisitMethodDeclaration(MethodDeclarationSyntax node)
-    {
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (IsIgnoreMember(node.Modifiers)) { return; }
-        foreach (var parameter in node.ParameterList?.Parameters)
-        {
-            var associationAttrSyntax = parameter.AttributeLists.GetAssociationAttributeSyntax();
-            if (associationAttrSyntax is not null)
-            {
-                var associationAttr = CreateAssociationAttribute(associationAttrSyntax);
-                relationships.AddAssociationFrom(node, parameter, associationAttr);
-            }
-        }
-        var modifiers = GetMemberModifiersText(node.Modifiers,
-                isInterfaceMember: node.Parent.IsKind(SyntaxKind.InterfaceDeclaration));
-        var name = node.Identifier.ToString();
-        var returnType = node.ReturnType.ToString();
-        var args = node.ParameterList.Parameters.Select(p => $"{p.Identifier}:{p.Type}");
 
-        WriteLine($"{modifiers}{name}({string.Join(", ", args)}) : {returnType}");
-    }
-
-    public override void VisitEnumMemberDeclaration(EnumMemberDeclarationSyntax node)
-    {
-        WriteLine($"{node.Identifier}{node.EqualsValue},");
-    }
-
-    public override void VisitEventFieldDeclaration(EventFieldDeclarationSyntax node)
-    {
-        if (IsIgnoreMember(node.Modifiers)) { return; }
-
-        var modifiers = GetMemberModifiersText(node.Modifiers,
-                isInterfaceMember: node.Parent.IsKind(SyntaxKind.InterfaceDeclaration));
-        var name = string.Join(",", node.Declaration.Variables.Select(v => v.Identifier));
-        var typeName = node.Declaration.Type.ToString();
-
-        WriteLine($"{modifiers} <<{node.EventKeyword}>> {name} : {typeName} ");
-    }
 
     public override void VisitGenericName(GenericNameSyntax node)
     {
@@ -393,112 +88,6 @@ public class ClassDiagramGenerator(
         if (nestingDepth == 1 && addPackageTags) return false;
         additionalTypeDeclarationNodes.Add(node);
         return true;
-    }
-
-    private void GenerateAdditionalTypeDeclarations()
-    {
-        for (int i = 0; i < additionalTypeDeclarationNodes.Count; i++)
-        {
-            SyntaxNode node = additionalTypeDeclarationNodes[i];
-            if (node is GenericNameSyntax genericNode)
-            {
-                if (createAssociation)
-                {
-                    GenerateAdditionalGenericTypeDeclaration(genericNode);
-                }
-                continue;
-            }
-            Visit(node);
-        }
-    }
-
-    private void GenerateAdditionalGenericTypeDeclaration(GenericNameSyntax genericNode)
-    {
-        var typename = TypeNameText.From(genericNode);
-        if (!types.Contains(typename.Identifier))
-        {
-            WriteLine($"class {typename.Identifier}{typename.TypeArguments} {{");
-            WriteLine("}");
-            types.Add(typename.Identifier);
-        }
-    }
-
-    private void GenerateRelationships()
-    {
-        foreach (var relationship in relationships)
-        {
-            WriteLine(relationship.ToString());
-        }
-    }
-    
-    public static string[] GenerateRelationships(RelationshipCollection relationshipCollection)
-    {
-        List<string> strings = new List<string>();
-        foreach (var relationship in relationshipCollection)
-        {
-            strings.AddRange(relationshipCollection.Select(r => r.ToString()));
-        }
-
-        return strings.ToArray();
-    }
-    
-    private void VisitFileScopedNamespaceDeclaration(FileScopedNamespaceDeclarationSyntax node, Action visitBase)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        var typeName = NamespaceNameText.From(node);
-
-        WriteLine($"package \"{typeName.Identifier}\" {{");
-        nestingDepth++;
-        visitBase();
-        nestingDepth--;
-        WriteLine("}");
-    }
-    
-    private void VisitNamespaceDeclaration(NamespaceDeclarationSyntax node, Action visitBase)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        var typeName = NamespaceNameText.From(node);
-
-        WriteLine($"package \"{typeName.Identifier}\"{{");
-        nestingDepth++;
-        visitBase();
-        nestingDepth--;
-        WriteLine("}");
-    }
-
-    private void VisitTypeDeclaration(TypeDeclarationSyntax node, Action visitBase)
-    {
-        if (attributeRequired && !node.AttributeLists.HasDiagramAttribute()) { return; }
-        if (node.AttributeLists.HasIgnoreAttribute()) { return; }
-        if (SkipInnerTypeDeclaration(node)) { return; }
-
-        relationships.AddInnerclassRelationFrom(node);
-        relationships.AddInheritanceFrom(node);
-
-        var modifiers = GetTypeModifiersText(node.Modifiers);
-        var keyword = (node.Modifiers.Any(SyntaxKind.AbstractKeyword) ? "abstract " : "")
-            + node.Keyword.ToString();
-
-        var typeName = TypeNameText.From(node);
-        var name = typeName.Identifier;
-        var typeParam = typeName.TypeArguments;
-        var type = $"{name}{typeParam}";
-
-        types.Add(name);
-
-        WriteLine($"{keyword} {type} {modifiers}{{");
-
-        nestingDepth++;
-        visitBase();
-        nestingDepth--;
-
-        WriteLine("}");
     }
 
     private static string GetTypeModifiersText(SyntaxTokenList modifiers)
